adicionando os arquivos

+++ projetoDjango/tasks/views.py
@@ -5,6 +5,7 @@ from .models import Task
 from django.contrib import messages
 from django.core.paginator import Paginator
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth import get_user_model
 
 @login_required
 def taskList(request):
@@ -12,9 +13,9 @@ def taskList(request):
     search = request.GET.get('search')
 
     if search:
-        tasks = Task.objects.filter(title__icontains=search)
+        tasks = Task.objects.filter(title__icontains=search, user=request.user)
     else:
-        tasks_list = Task.objects.all().order_by('-created_at')
+        tasks_list = Task.objects.all().order_by('-created_at').filter(user=request.user)
         paginator = Paginator(tasks_list, 3)
         page = request.GET.get('page')
         tasks = paginator.get_page(page)
@@ -24,7 +25,11 @@ def taskList(request):
 @login_required
 def taskView(request, id):
     task = get_object_or_404(Task, pk=id)
-    return render(request, 'tasks/task.html', {'task': task})
+    if request.user != task.user:
+        messages.info(request, 'Você não tem permissão para visualizar essa tarefa!')
+        return redirect('/')
+    else:
+        return render(request, 'tasks/task.html', {'task': task})
 
 @login_required
 def newTask(request):
@@ -33,6 +38,7 @@ def newTask(request):
         if form.is_valid():
             task = form.save(commit=False)
             task.done = 'doing'
+            task.user = request.user
             task.save()
             return redirect('/')
     else:
@@ -42,25 +48,33 @@ def newTask(request):
 @login_required
 def editTask(request, id):
     task = get_object_or_404(Task, pk=id)
-    form = TaskForm(instance=task)
-    if(request.method == 'POST'):
-        form = TaskForm(request.POST, instance=task)
-        if(form.is_valid()):
-            task.save()
-            return redirect('/')
+    if task.user != request.user:
+        messages.info(request, 'Você não tem permissão para editar essa tarefa!')
+        return redirect('/')
+    else:
+        form = TaskForm(instance=task)
+        if(request.method == 'POST'):
+            form = TaskForm(request.POST, instance=task)
+            if(form.is_valid()):
+                task.save()
+                return redirect('/')
+            else:
+                return render(request, 'tasks/edittask.html', {'form': form, 'task': task})
         else:
             return render(request, 'tasks/edittask.html', {'form': form, 'task': task})
-    else:
-        return render(request, 'tasks/edittask.html', {'form': form, 'task': task})
 
 @login_required
 def deleteTask(request, id):
     task = get_object_or_404(Task, pk=id)
-    task.delete()
+    if task.user != request.user:
+        messages.info(request, 'Você não tem permissão para excluir essa tarefa!')
+        return redirect('/')
+    else:
+        task.delete()
 
-    messages.info(request, 'Tarefa deletada com sucesso.')
+        messages.info(request, 'Tarefa deletada com sucesso.')
 
-    return redirect('/')
+        return redirect('/')
 
 def helloWorld(request):
     return HttpResponse('Hello World!')
